refactor: split function call resolution
