+++ src/types/checker/functions/resolution/mod.rs
@@ -1,627 +1,13 @@
 //! Purpose:
-//! Resolves function call targets and callable signatures for the checker.
-//! Combines declared functions, builtin signatures, specialization, and callable metadata into one lookup path.
+//! Organizes named-function resolution, signature publication, and specialization.
 //!
 //! Called from:
-//! - `crate::types::checker::functions`
+//! - `crate::types::checker::functions`.
 //!
 //! Key details:
-//! - Resolution must match name-resolver canonicalization and PHP builtin fallback rules.
+//! - Direct and pre-normalized calls converge on the same resolved-signature validator.
 
+mod call;
+mod resolved;
 mod signature;
 mod specialization;
-
-use crate::errors::CompileError;
-use crate::parser::ast::{Expr, ExprKind};
-use crate::types::{FunctionSig, PhpType, TypeEnv};
-
-use super::super::Checker;
-
-impl Checker {
-    /// Determines the element type for a variadic parameter's container.
-    ///
-    /// If the element type is `Iterable`, returns `Mixed` since iterables can
-    /// hold heterogeneous values at runtime. Otherwise returns the element type unchanged.
-    fn variadic_container_elem_ty(elem_ty: PhpType) -> PhpType {
-        if matches!(elem_ty, PhpType::Iterable) {
-            PhpType::Mixed
-        } else {
-            elem_ty
-        }
-    }
-
-    /// Returns true if any named argument in `args` does not correspond to a
-    /// declared regular (non-variadic) parameter in `regular_params`.
-    ///
-    /// Used to decide whether a variadic parameter should receive type `Iterable`
-    /// (when unknown named args may flow into it) or `Array<elem>` (when all
-    /// named args are known).
-    fn has_unknown_named_variadic_arg(args: &[Expr], regular_params: &[String]) -> bool {
-        args.iter().any(|arg| {
-            matches!(
-                &arg.kind,
-                ExprKind::NamedArg { name, .. } if !regular_params.iter().any(|param| param == name)
-            )
-        })
-    }
-
-    /// Resolves a function call by name, normalizes arguments (named, spread),
-    /// validates arity and types, applies specialization if needed, and returns
-    /// the function's return type.
-    ///
-    /// Handles three lookup paths:
-    /// - Declared functions already in `self.functions`
-    /// - Function variant groups (which may emit a provisional signature)
-    /// - Forward declarations from `fn_decls` (where type hints are resolved here)
-    ///
-    /// Emits a deprecation warning if the function is deprecated.
-    /// Re-specializes resolved function parameters when the call site demands it.
-    /// Builds the diagnostic for a call that resolved to no declared function, builtin signature,
-    /// variant group, or extern. When `name` is a procedural date/time alias (which the name
-    /// resolver desugars only at its supported arities), reaching this point means the call's
-    /// argument count was out of range, so report a precise arity error — matching
-    /// `function_exists()`, which already recognizes these names — instead of the misleading
-    /// "Undefined function". Non-alias names keep the plain "Undefined function" diagnostic.
-    fn unresolved_function_call_error(&self, name: &str, span: crate::span::Span) -> CompileError {
-        if let Some((min, max)) = crate::name_resolver::date_procedural_alias_arity(name) {
-            let bare = name.rsplit('\\').next().unwrap_or(name);
-            let message = if min == max {
-                format!(
-                    "{}() takes exactly {} argument{}",
-                    bare,
-                    min,
-                    if min == 1 { "" } else { "s" }
-                )
-            } else if max == min + 1 {
-                format!("{}() takes {} or {} arguments", bare, min, max)
-            } else {
-                format!("{}() takes {} to {} arguments", bare, min, max)
-            };
-            return CompileError::new(span, &message);
-        }
-        // Under `--strict-php`, a name that would resolve to an extension builtin
-        // lands here as undefined. Name the disabled extension so users understand
-        // why a working non-strict program stopped compiling.
-        let bare = name.rsplit('\\').next().unwrap_or(name);
-        let key = crate::names::php_symbol_key(bare);
-        if crate::types::checker::builtins::strict_php_hidden_builtin(&key) {
-            return CompileError::new(
-                span,
-                &format!(
-                    "Undefined function: {} ({}() exists as an elephc extension; it is disabled by --strict-php)",
-                    name, key
-                ),
-            );
-        }
-        CompileError::new(span, &format!("Undefined function: {}", name))
-    }
-
-    /// Checks a function call, including externs, declared functions, variants,
-    /// builtins, call-argument normalization, and return-type inference.
-    pub fn check_function_call(
-        &mut self,
-        name: &str,
-        args: &[Expr],
-        span: crate::span::Span,
-        caller_env: &TypeEnv,
-    ) -> Result<PhpType, CompileError> {
-        if let Some(extern_name) = self.canonical_extern_function_name_folded(name) {
-            return self.check_extern_function_call(&extern_name, args, span, caller_env);
-        }
-        let canonical_name = self
-            .canonical_function_name_folded(name)
-            .unwrap_or_else(|| name.to_string());
-        let name = canonical_name.as_str();
-
-        if let Some(mut sig) = self.functions.get(name).cloned() {
-            if let Some(reason) = sig.deprecation.as_deref() {
-                let message = if reason.is_empty() {
-                    format!("Call to deprecated function: {}()", name)
-                } else {
-                    format!("Call to deprecated function: {}() — {}", name, reason)
-                };
-                self.warnings
-                    .push(crate::errors::CompileWarning::new(span, &message));
-            }
-            let mut effective_sig =
-                Self::callable_sig_for_declared_params(&sig, &sig.declared_params);
-            let normalized_args = self.normalize_named_call_args(
-                &effective_sig,
-                args,
-                span,
-                &format!("Function '{}'", name),
-                caller_env,
-            )?;
-            if self.respecialize_resolved_function_params_if_needed(
-                name,
-                &normalized_args,
-                caller_env,
-            )? {
-                sig = self.functions.get(name).cloned().ok_or_else(|| {
-                    CompileError::new(span, &format!("Undefined function: {}", name))
-                })?;
-                effective_sig =
-                    Self::callable_sig_for_declared_params(&sig, &sig.declared_params);
-            }
-            return self.check_normalized_resolved_function_call(
-                name,
-                &sig,
-                &effective_sig,
-                &normalized_args,
-                span,
-                caller_env,
-            );
-        }
-
-        if self.function_variant_groups.contains_key(name) {
-            let ensure_error = self
-                .ensure_function_variant_group_signature(name, span)
-                .err();
-            if let Some(error) = ensure_error.as_ref() {
-                if !self.functions.contains_key(name) {
-                    return Err(error.clone());
-                }
-            }
-            let provisional_sig = self.functions.get(name).cloned();
-            let result = self.check_function_call(name, args, span, caller_env);
-            if let Some(error) = ensure_error {
-                if provisional_sig == self.functions.get(name).cloned() {
-                    return Err(error);
-                }
-            }
-            return result;
-        }
-
-        if self.eval_barrier_active {
-            for arg in args {
-                self.infer_type(arg, caller_env)?;
-            }
-            return Ok(PhpType::Mixed);
-        }
-
-        let decl = self
-            .fn_decls
-            .get(name)
-            .cloned()
-            .ok_or_else(|| self.unresolved_function_call_error(name, span))?;
-        let normalization_sig = FunctionSig {
-            params: decl
-                .params
-                .iter()
-                .enumerate()
-                .map(|(idx, param_name)| {
-                    let ty = decl
-                        .param_types
-                        .get(idx)
-                        .and_then(|type_ann| type_ann.as_ref())
-                        .and_then(|type_ann| self.resolve_type_expr(type_ann, decl.span).ok())
-                        .unwrap_or(PhpType::Int);
-                    (param_name.clone(), ty)
-                })
-                .chain(
-                    decl.variadic
-                        .iter()
-                        .cloned()
-                        .map(|name| {
-                            let elem_ty = if decl.variadic_by_ref {
-                                PhpType::Mixed
-                            } else {
-                                PhpType::Int
-                            };
-                            (name, PhpType::Array(Box::new(elem_ty)))
-                        }),
-                )
-                .collect(),
-            param_type_exprs: decl
-                .param_types
-                .iter()
-                .cloned()
-                .chain(decl.variadic.iter().map(|_| decl.variadic_type.clone()))
-                .collect(),
-            param_attributes: decl.param_attributes.clone(),
-            defaults: decl.defaults.clone(),
-            return_type: PhpType::Int,
-            declared_return: decl.return_type.is_some(),
-            by_ref_return: false,
-            ref_params: decl.ref_params.clone(),
-            declared_params: decl
-                .param_types
-                .iter()
-                .map(|type_ann| type_ann.is_some())
-                .chain(decl.variadic.iter().map(|_| decl.variadic_type.is_some()))
-                .collect(),
-            variadic: decl.variadic.clone(),
-            deprecation: None,
-        };
-        let normalized_args = self.normalize_named_call_args(
-            &normalization_sig,
-            args,
-            span,
-            &format!("Function '{}'", name),
-            caller_env,
-        )?;
-        let args = normalized_args.as_slice();
-        let effective_arg_count = args
-            .iter()
-            .filter(|a| !matches!(a.kind, ExprKind::Spread(_)))
-            .count();
-        let has_spread = args.iter().any(|a| matches!(a.kind, ExprKind::Spread(_)));
-        let required = decl.defaults.iter().filter(|d| d.is_none()).count();
-        if decl.ref_params.iter().any(|is_ref| *is_ref) && has_spread {
-            return Err(CompileError::new(
-                span,
-                &format!(
-                    "Function '{}' cannot be invoked with spread arguments when it has pass-by-reference parameters",
-                    name
-                ),
-            ));
-        }
-        if !has_spread {
-            if decl.variadic.is_some() {
-                if effective_arg_count < required {
-                    return Err(CompileError::new(
-                        span,
-                        &format!(
-                            "Function '{}' expects at least {} arguments, got {}",
-                            name, required, effective_arg_count
-                        ),
-                    ));
-                }
-            } else if effective_arg_count < required || effective_arg_count > decl.params.len() {
-                return Err(CompileError::new(
-                    span,
-                    &format!(
-                        "Function '{}' expects {} arguments, got {}",
-                        name,
-                        Self::format_fixed_or_range_arity(required, decl.params.len()),
-                        effective_arg_count
-                    ),
-                ));
-            }
-        }
-
-        let mut param_types = Vec::new();
-        let mut arg_idx = 0;
-        for arg in args {
-            let ty = self.infer_type(arg, caller_env)?;
-            if let ExprKind::Spread(_) = &arg.kind {
-                for i in arg_idx..decl.params.len() {
-                    if let Some(type_ann) = decl.param_types.get(i).and_then(|t| t.as_ref()) {
-                        let declared_ty = self.resolve_declared_param_type_hint(
-                            type_ann,
-                            decl.span,
-                            &format!("Function '{}' parameter ${}", name, decl.params[i]),
-                        )?;
-                        self.require_compatible_arg_type(
-                            &declared_ty,
-                            &ty,
-                            arg.span,
-                            &format!("Function '{}' parameter ${}", name, decl.params[i]),
-                        )?;
-                        param_types.push((decl.params[i].clone(), declared_ty));
-                    } else if matches!(ty, PhpType::Never) {
-                        let param_ty = decl
-                            .defaults
-                            .get(i)
-                            .and_then(|default| default.as_ref())
-                            .map(|default| self.infer_type(default, caller_env))
-                            .transpose()?
-                            .unwrap_or(PhpType::Int);
-                        param_types.push((decl.params[i].clone(), param_ty));
-                    } else {
-                        param_types.push((decl.params[i].clone(), ty.clone()));
-                    }
-                }
-                arg_idx = decl.params.len();
-            } else if arg_idx < decl.params.len() {
-                if ty == PhpType::Callable {
-                    if let Some(sig) = self.resolve_expr_callable_sig(arg, caller_env)? {
-                        self.callable_param_sigs.insert(
-                            (name.to_string(), decl.params[arg_idx].clone()),
-                            sig,
-                        );
-                    }
-                }
-                if is_callable_array_type(&ty) {
-                    if let Some(sig) = self.resolve_expr_callable_array_sig(arg, caller_env)? {
-                        self.callable_param_sigs.insert(
-                            (name.to_string(), decl.params[arg_idx].clone()),
-                            sig,
-                        );
-                    }
-                }
-                if decl.ref_params.get(arg_idx).copied().unwrap_or(false)
-                    && !self.is_by_ref_argument_lvalue(arg, caller_env)?
-                {
-                    let param_name = decl
-                        .params
-                        .get(arg_idx)
-                        .map(String::as_str)
-                        .unwrap_or("arg");
-                    return Err(CompileError::new(
-                        arg.span,
-                        &format!(
-                            "Function '{}' parameter ${} must be passed a variable",
-                            name, param_name
-                        ),
-                    ));
-                }
-                if let Some(type_ann) = decl.param_types.get(arg_idx).and_then(|t| t.as_ref()) {
-                    let param_name = decl
-                        .params
-                        .get(arg_idx)
-                        .map(String::as_str)
-                        .unwrap_or("arg");
-                    let declared_ty = self.resolve_declared_param_type_hint(
-                        type_ann,
-                        decl.span,
-                        &format!("Function '{}' parameter ${}", name, param_name),
-                    )?;
-                    if decl.ref_params.get(arg_idx).copied().unwrap_or(false) {
-                        self.require_boxed_by_ref_storage(
-                            &declared_ty,
-                            &ty,
-                            arg.span,
-                            &format!("Function '{}' parameter ${}", name, param_name),
-                        )?;
-                    }
-                    self.require_compatible_arg_type(
-                        &declared_ty,
-                        &ty,
-                        arg.span,
-                        &format!("Function '{}' parameter ${}", name, param_name),
-                    )?;
-                    let specialized_ty =
-                        Self::specialize_generic_array_param_hint(&declared_ty, &ty);
-                    param_types.push((decl.params[arg_idx].clone(), specialized_ty));
-                    arg_idx += 1;
-                    continue;
-                }
-                param_types.push((decl.params[arg_idx].clone(), ty));
-                arg_idx += 1;
-            } else {
-                // Argument collected into the variadic parameter: enforce its declared element
-                // type (`int ...$xs`) against every passed argument, matching PHP.
-                if let Some(declared) = &decl.variadic_type {
-                    let vname = decl.variadic.as_deref().unwrap_or_default();
-                    let elem_ty = self.resolve_declared_param_type_hint(
-                        declared,
-                        decl.span,
-                        &format!("Function '{}' variadic parameter ${}", name, vname),
-                    )?;
-                    self.require_compatible_arg_type(
-                        &elem_ty,
-                        &ty,
-                        arg.span,
-                        &format!("Function '{}' variadic parameter ${}", name, vname),
-                    )?;
-                }
-                arg_idx += 1;
-            }
-        }
-
-        for i in arg_idx..decl.params.len() {
-            if let Some(default_expr) = &decl.defaults[i] {
-                if let Some(type_ann) = decl.param_types.get(i).and_then(|t| t.as_ref()) {
-                    let declared_ty = self.resolve_declared_param_type_hint(
-                        type_ann,
-                        decl.span,
-                        &format!("Function '{}' parameter ${}", name, decl.params[i]),
-                    )?;
-                    let default_ty = self.infer_type(default_expr, caller_env)?;
-                    self.require_compatible_arg_type(
-                        &declared_ty,
-                        &default_ty,
-                        default_expr.span,
-                        &format!("Function '{}' parameter ${}", name, decl.params[i]),
-                    )?;
-                    param_types.push((decl.params[i].clone(), declared_ty));
-                } else {
-                    let ty = self.infer_type(default_expr, caller_env)?;
-                    param_types.push((decl.params[i].clone(), ty));
-                }
-            }
-        }
-
-        if let Some(ref vp) = decl.variadic {
-            if decl.variadic_by_ref {
-                param_types.push((vp.clone(), PhpType::Array(Box::new(PhpType::Mixed))));
-            } else if let Some(declared) = &decl.variadic_type {
-                // A declared element type (`int ...$xs`) constrains every collected argument;
-                // it takes precedence over inference so call validation enforces the hint.
-                let elem_ty = self.resolve_declared_param_type_hint(
-                    declared,
-                    decl.span,
-                    &format!("Variadic parameter ${}", vp),
-                )?;
-                param_types.push((vp.clone(), PhpType::Array(Box::new(elem_ty))));
-            } else if Self::has_unknown_named_variadic_arg(args, &decl.params) {
-                param_types.push((vp.clone(), PhpType::Iterable));
-            } else {
-                let variadic_elem_ty = if args.len() > decl.params.len() {
-                    self.infer_type(&args[decl.params.len()], caller_env)
-                        .unwrap_or(PhpType::Int)
-                } else {
-                    PhpType::Int
-                };
-                let variadic_elem_ty = Self::variadic_container_elem_ty(variadic_elem_ty);
-                param_types.push((vp.clone(), PhpType::Array(Box::new(variadic_elem_ty))));
-            }
-        }
-
-        self.resolve_function_signature(name, &decl, param_types)
-    }
-
-    /// Checks a function call where arguments are already normalized (named,
-    /// spread, and order resolved) against a function already present in
-    /// `self.functions`.
-    ///
-    /// Used by variant-group resolution and other paths that have a pre-resolved
-    /// signature and do not need re-specialization or deprecation warnings.
-    pub(crate) fn check_function_call_pre_normalized(
-        &mut self,
-        name: &str,
-        normalized_args: &[Expr],
-        span: crate::span::Span,
-        caller_env: &TypeEnv,
-    ) -> Result<PhpType, CompileError> {
-        let sig = self
-            .functions
-            .get(name)
-            .cloned()
-            .ok_or_else(|| self.unresolved_function_call_error(name, span))?;
-        let effective_sig = Self::callable_sig_for_declared_params(&sig, &sig.declared_params);
-        self.check_normalized_resolved_function_call(
-            name,
-            &sig,
-            &effective_sig,
-            normalized_args,
-            span,
-            caller_env,
-        )
-    }
-
-    /// Validates a resolved, normalized function call against its effective
-    /// signature: arity constraints, by-ref argument validation, and type
-    /// compatibility for each argument position (regular and variadic).
-    ///
-    /// Does **not** re-specialize or check deprecation — those are handled by
-    /// the caller before dispatching here. Returns the signature's return type
-    /// on success.
-    fn check_normalized_resolved_function_call(
-        &mut self,
-        name: &str,
-        sig: &FunctionSig,
-        effective_sig: &FunctionSig,
-        args: &[Expr],
-        span: crate::span::Span,
-        caller_env: &TypeEnv,
-    ) -> Result<PhpType, CompileError> {
-        let effective_arg_count = args
-            .iter()
-            .filter(|a| !matches!(a.kind, ExprKind::Spread(_)))
-            .count();
-        let has_spread = args.iter().any(|a| matches!(a.kind, ExprKind::Spread(_)));
-        let required = effective_sig.defaults.iter().filter(|d| d.is_none()).count();
-        if effective_sig.ref_params.iter().any(|is_ref| *is_ref) && has_spread {
-            return Err(CompileError::new(
-                span,
-                &format!(
-                    "Function '{}' cannot be invoked with spread arguments when it has pass-by-reference parameters",
-                    name
-                ),
-            ));
-        }
-        if !has_spread {
-            if effective_sig.variadic.is_some() {
-                if effective_arg_count < required {
-                    return Err(CompileError::new(
-                        span,
-                        &format!(
-                            "Function '{}' expects at least {} arguments, got {}",
-                            name, required, effective_arg_count
-                        ),
-                    ));
-                }
-            } else if effective_arg_count < required
-                || effective_arg_count > effective_sig.params.len()
-            {
-                return Err(CompileError::new(
-                    span,
-                    &format!(
-                        "Function '{}' expects {} arguments, got {}",
-                        name,
-                        Self::format_fixed_or_range_arity(required, effective_sig.params.len()),
-                        effective_arg_count
-                    ),
-                ));
-            }
-        }
-        let regular_param_count = if effective_sig.variadic.is_some() {
-            effective_sig.params.len().saturating_sub(1)
-        } else {
-            effective_sig.params.len()
-        };
-        let variadic_elem_ty = effective_sig.variadic.as_ref().and_then(|_| {
-            effective_sig.params.last().and_then(|(_, ty)| match ty {
-                PhpType::Array(elem) => Some((**elem).clone()),
-                _ => None,
-            })
-        });
-        let mut param_idx = 0usize;
-        for arg in args {
-            let actual_ty = self.infer_type(arg, caller_env)?;
-            if matches!(arg.kind, ExprKind::Spread(_)) {
-                continue;
-            }
-            if param_idx < regular_param_count {
-                if effective_sig
-                    .ref_params
-                    .get(param_idx)
-                    .copied()
-                    .unwrap_or(false)
-                    && !self.is_by_ref_argument_lvalue(arg, caller_env)?
-                {
-                    let param_name = effective_sig
-                        .params
-                        .get(param_idx)
-                        .map(|(name, _)| name.as_str())
-                        .unwrap_or("arg");
-                    return Err(CompileError::new(
-                        arg.span,
-                        &format!(
-                            "Function '{}' parameter ${} must be passed a variable",
-                            name, param_name
-                        ),
-                    ));
-                }
-                if let Some((param_name, expected_ty)) = effective_sig.params.get(param_idx) {
-                    if effective_sig
-                        .declared_params
-                        .get(param_idx)
-                        .copied()
-                        .unwrap_or(false)
-                        && effective_sig
-                            .ref_params
-                            .get(param_idx)
-                            .copied()
-                            .unwrap_or(false)
-                    {
-                        self.require_boxed_by_ref_storage(
-                            expected_ty,
-                            &actual_ty,
-                            arg.span,
-                            &format!("Function '{}' parameter ${}", name, param_name),
-                        )?;
-                    }
-                    self.require_compatible_arg_type(
-                        expected_ty,
-                        &actual_ty,
-                        arg.span,
-                        &format!("Function '{}' parameter ${}", name, param_name),
-                    )?;
-                }
-            } else if let (Some(vname), Some(expected_ty)) =
-                (effective_sig.variadic.as_ref(), variadic_elem_ty.as_ref())
-            {
-                self.require_compatible_arg_type(
-                    expected_ty,
-                    &actual_ty,
-                    arg.span,
-                    &format!("Function '{}' variadic parameter ${}", name, vname),
-                )?;
-            }
-            param_idx += 1;
-        }
-        Ok(sig.return_type.clone())
-    }
-}
-
-/// Returns true when a call argument is an array whose elements are callable descriptors.
-fn is_callable_array_type(ty: &PhpType) -> bool {
-    match ty {
-        PhpType::Array(elem_ty) => elem_ty.as_ref() == &PhpType::Callable,
-        PhpType::AssocArray { value, .. } => value.as_ref() == &PhpType::Callable,
-        _ => false,
-    }
-}
